etl-system/requirements.txt
cassandra-driver==3.29.2
cffi==1.17.1
click==8.1.8
confluent-kafka==2.8.0
cryptography==44.0.0
geomet==0.2.1.post1
greenlet==3.1.1
numpy==1.24.3  # Updated version
pandas==2.0.3  # Updated version
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.2